Moved sim tod reading to separate script.
